+++ src/helper/color/rgbToRGBA.ts
@@ -1,4 +1,9 @@
-export default function rgbToRGBA(color: string, alpha: number) {
+import { Duck } from '../../index';
+
+export default function rgbToRGBA(
+	color: string,
+	alpha: Duck.Helper.AlphaRange
+) {
 	let new_col = '';
 	new_col = color.replace(/rgb/i, 'rgba');
 	new_col = new_col.replace(/\)/i, `,${alpha})`);