+++ src/index.ts
@@ -25,6 +25,7 @@ import SoundClass from './core/sound/sound';
 import DuckStorageClass from './core/storage/storage';
 import OnceClass from './base/once';
 import RenderClass from './base/render';
+import ButtonClass from './core/interactive/button';
 
 // main
 
@@ -47,10 +48,12 @@ export namespace Duck {
 		}
 
 		export type Group<t extends Duck.Group.StackItem> = GroupClass<t>;
+
 		export type Input = InputClass;
 
 		export namespace Interactive {
 			export type Text = TextClass;
+			export type Button = ButtonClass;
 		}
 
 		export namespace Lights {
@@ -314,7 +317,77 @@ export namespace Duck {
 		}
 	}
 
-	namespace Helper {
+	export namespace TouchInput {
+		export type Pos =
+			| 'TOP-LEFT'
+			| 'TOP-RIGHT'
+			| 'BOTTOM-LEFT'
+			| 'BOTTOM-RIGHT';
+
+		export interface Controller {
+			up: boolean;
+			down: boolean;
+			left: boolean;
+			right: boolean;
+		}
+
+		export interface Controls {
+			up?: boolean;
+			down?: boolean;
+			left?: boolean;
+			right?: boolean;
+		}
+
+		export interface OnTouch {
+			up?: (e: TouchEvent) => void;
+			down?: (e: TouchEvent) => void;
+			left?: (e: TouchEvent) => void;
+			right?: (e: TouchEvent) => void;
+		}
+
+		export interface Styles {
+			buttonW: string;
+			buttonH: string;
+			border?: string;
+			borderRadius?: string;
+			zIndex?: Helper.DefaultValue<undefined, number>;
+			spacing?: Helper.DefaultValue<undefined, number>;
+			textColor?: Helper.DefaultValue<undefined, string>;
+			bgColor?: Helper.DefaultValue<undefined, string>;
+			up?: {
+				innerHTML: string;
+				w?: string;
+				h?: string;
+				textColor?: string;
+				bgColor?: string;
+			};
+			down?: {
+				innerHTML: string;
+				w?: string;
+				h?: string;
+				textColor?: string;
+				bgColor?: string;
+			};
+			left?: {
+				innerHTML: string;
+				w?: string;
+				h?: string;
+				textColor?: string;
+				bgColor?: string;
+			};
+			right?: {
+				innerHTML: string;
+				w?: string;
+				h?: string;
+				textColor?: string;
+				bgColor?: string;
+			};
+		}
+
+		export type ButtonType = 'up' | 'down' | 'left' | 'right';
+	}
+
+	export namespace Helper {
 		type ArrayLengthMutationKeys =
 			| 'splice'
 			| 'push'
@@ -331,6 +404,120 @@ export namespace Duck {
 			T,
 			Exclude<keyof T, ArrayLengthMutationKeys>
 		> & { [Symbol.iterator]: () => IterableIterator<ArrayItems<T>> };
+
+		export type NonNullable<T> = T extends null | undefined ? never : T;
+		export type DefaultValue<
+			Type,
+			Default,
+			ExtraRule = null
+		> = Type extends null | undefined | false | 0 | ExtraRule | ''
+			? Default
+			: Type;
+
+		//#region Long Type
+		export type AlphaRange =
+			| 0
+			| 0.01
+			| 0.02
+			| 0.03
+			| 0.04
+			| 0.05
+			| 0.06
+			| 0.07
+			| 0.08
+			| 0.09
+			| 0.1
+			| 0.11
+			| 0.12
+			| 0.13
+			| 0.14
+			| 0.15
+			| 0.16
+			| 0.17
+			| 0.18
+			| 0.19
+			| 0.2
+			| 0.21
+			| 0.22
+			| 0.23
+			| 0.24
+			| 0.25
+			| 0.26
+			| 0.27
+			| 0.28
+			| 0.29
+			| 0.3
+			| 0.31
+			| 0.32
+			| 0.33
+			| 0.34
+			| 0.35
+			| 0.36
+			| 0.37
+			| 0.38
+			| 0.39
+			| 0.4
+			| 0.41
+			| 0.42
+			| 0.43
+			| 0.44
+			| 0.45
+			| 0.46
+			| 0.47
+			| 0.48
+			| 0.49
+			| 0.5
+			| 0.51
+			| 0.52
+			| 0.53
+			| 0.54
+			| 0.55
+			| 0.56
+			| 0.57
+			| 0.58
+			| 0.59
+			| 0.6
+			| 0.61
+			| 0.62
+			| 0.63
+			| 0.64
+			| 0.65
+			| 0.66
+			| 0.67
+			| 0.68
+			| 0.69
+			| 0.7
+			| 0.71
+			| 0.72
+			| 0.73
+			| 0.74
+			| 0.75
+			| 0.76
+			| 0.77
+			| 0.78
+			| 0.79
+			| 0.8
+			| 0.81
+			| 0.82
+			| 0.83
+			| 0.84
+			| 0.85
+			| 0.86
+			| 0.87
+			| 0.88
+			| 0.89
+			| 0.9
+			| 0.91
+			| 0.92
+			| 0.93
+			| 0.94
+			| 0.95
+			| 0.96
+			| 0.97
+			| 0.98
+			| 0.99
+			| 1;
+		//#endregion Long Type
 	}
 }